Better Rtten Flesh
The Rotten Flesh is burned in a furnace to make a fabric mod for leather. It also restores the recipe for the Enchanted Golden Apple, which can be crafted with 8 gold nuggets and 1 apple, or upgraded to an enchanted golden apple with 1 ordinary golden apple plus 7 gold nuggets and 1 gold ingot. Plus, there's more hidden practical recipes waiting to be discovered